wip: start migrating merkle-tree to ECLAIR
