+++ manta-crypto/src/merkle_tree/forest.rs
@@ -317,6 +317,7 @@ where
 
     #[inline]
     fn prove(&self, item: &Self::Item) -> Option<MembershipProof<Self::Model>> {
+        /* TODO:
         let tree = self.forest.get_tree(item);
         Some(MembershipProof::new(
             tree.path(
@@ -326,6 +327,8 @@ where
             .ok()?,
             tree.root().clone(),
         ))
+        */
+        todo!()
     }
 
     #[inline]
