+++ manta-crypto/src/merkle_tree/inner_tree.rs
@@ -16,13 +16,9 @@
 
 //! Inner Digest Tree
 
-// TODO: Figure out how we want to expose the meaning of `InnerNode` coordinates. Should we share
-//       some of it, to reduce potential duplication?
-// TODO: We should probably move `InnerNode` and its related `struct`s to `merkle_tree::node`.
-
 use crate::merkle_tree::{
     path::{CurrentInnerPath, InnerPath},
-    path_length, Configuration, InnerDigest, Node, Parameters, Parity,
+    Configuration, InnerDigest, InnerNode, InnerNodeIter, Node, Parameters, Parity,
 };
 use alloc::collections::btree_map;
 use core::{fmt::Debug, hash::Hash, iter::FusedIterator, marker::PhantomData, ops::Index};
@@ -32,172 +28,6 @@ use manta_util::serde::{Deserialize, Serialize};
 
 #[cfg(feature = "std")]
 use std::{collections::hash_map, hash::BuildHasher};
-
-/// Inner Tree Node
-#[cfg_attr(
-    feature = "serde",
-    derive(Deserialize, Serialize),
-    serde(crate = "manta_util::serde", deny_unknown_fields)
-)]
-#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
-pub struct InnerNode {
-    /// Depth
-    depth: usize,
-
-    /// Index
-    index: Node,
-}
-
-impl InnerNode {
-    /// Builds a new [`InnerNode`] from `depth` and `index`.
-    #[inline]
-    const fn new(depth: usize, index: Node) -> Self {
-        Self { depth, index }
-    }
-
-    /// Builds an [`InnerNode`] as the parent of a `leaf_index`, returning `None` if the
-    /// parent of `leaf_index` is the root and not an inner node.
-    #[inline]
-    pub fn from_leaf<C>(leaf_index: Node) -> Option<Self>
-    where
-        C: Configuration + ?Sized,
-    {
-        Self::new(path_length::<C>(), leaf_index).parent()
-    }
-
-    /// Returns the [`Parity`] of this inner node.
-    #[inline]
-    pub const fn parity(&self) -> Parity {
-        self.index.parity()
-    }
-
-    /// Returns `true` if this inner node has left parity.
-    #[inline]
-    pub const fn is_left(&self) -> bool {
-        self.parity().is_left()
-    }
-
-    /// Returns `true` if this inner node has right parity.
-    #[inline]
-    pub const fn is_right(&self) -> bool {
-        self.parity().is_right()
-    }
-
-    /// Returns the [`InnerNode`] which is the sibling of `self`.
-    #[inline]
-    #[must_use]
-    pub const fn sibling(&self) -> Self {
-        Self::new(self.depth, self.index.sibling())
-    }
-
-    /// Returns the parent [`InnerNode`] of this inner node.
-    #[inline]
-    pub const fn parent(&self) -> Option<Self> {
-        match self.depth.checked_sub(1) {
-            Some(depth) => Some(Self::new(depth, self.index.parent())),
-            _ => None,
-        }
-    }
-
-    /// Converts `self` into its parent, if the parent exists, returning the parent [`InnerNode`].
-    #[inline]
-    pub fn into_parent(&mut self) -> Option<Self> {
-        match self.parent() {
-            Some(parent) => {
-                *self = parent;
-                Some(*self)
-            }
-            _ => None,
-        }
-    }
-
-    /// Returns an iterator over `self` and its parents.
-    #[inline]
-    pub const fn iter(&self) -> InnerNodeIter {
-        InnerNodeIter::new(Some(*self))
-    }
-
-    /// Computes the starting index for the given `self.depth` in the tree.
-    #[inline]
-    const fn depth_starting_index(&self) -> usize {
-        (1 << (self.depth + 1)) - 1
-    }
-
-    /// Computes an [`InnerMap`] index for the coordinates represented by `self`.
-    #[inline]
-    pub const fn map_index(&self) -> usize {
-        self.depth_starting_index() + self.index.0
-    }
-}
-
-impl From<InnerNode> for Node {
-    #[inline]
-    fn from(inner_node: InnerNode) -> Node {
-        inner_node.index
-    }
-}
-
-/// Inner Node Iterator
-///
-/// An iterator over the parents of an [`InnerNode`], including the node itself.
-///
-/// This `struct` is created by the [`iter`](InnerNode::iter) method on [`InnerNode`].
-/// See its documentation for more.
-#[cfg_attr(
-    feature = "serde",
-    derive(Deserialize, Serialize),
-    serde(crate = "manta_util::serde", deny_unknown_fields)
-)]
-#[derive(Clone, Copy, Debug, Default, Eq, Hash, PartialEq)]
-pub struct InnerNodeIter {
-    /// Current Node
-    node: Option<InnerNode>,
-}
-
-impl InnerNodeIter {
-    /// Builds a new [`InnerNodeIter`] from `node`.
-    #[inline]
-    const fn new(node: Option<InnerNode>) -> Self {
-        Self { node }
-    }
-
-    /// Builds a new [`InnerNodeIter`] iterator over the parents of `leaf_index`.
-    #[inline]
-    pub fn from_leaf<C>(leaf_index: Node) -> Self
-    where
-        C: Configuration + ?Sized,
-    {
-        Self::new(InnerNode::from_leaf::<C>(leaf_index))
-    }
-
-    /// Returns `true` if the iterator has completed.
-    #[inline]
-    pub const fn is_done(&self) -> bool {
-        self.node.is_none()
-    }
-}
-
-// TODO: Add all methods which can be optimized.
-impl Iterator for InnerNodeIter {
-    type Item = InnerNode;
-
-    #[inline]
-    fn next(&mut self) -> Option<Self::Item> {
-        let node = self.node.take()?;
-        self.node = node.parent();
-        Some(node)
-    }
-
-    #[inline]
-    fn size_hint(&self) -> (usize, Option<usize>) {
-        let len = self.node.map_or(0, move |n| n.depth + 1);
-        (len, Some(len))
-    }
-}
-
-impl ExactSizeIterator for InnerNodeIter {}
-
-impl FusedIterator for InnerNodeIter {}
 
 /// [`InnerTree`] Map Backend
 pub trait InnerMap<C>
