+++ manta-crypto/src/merkle_tree/mod.rs
@@ -19,7 +19,6 @@
 // NOTE: clippy false-positive: https://github.com/rust-lang/rust-clippy/pull/8869
 #![allow(clippy::derive_partial_eq_without_eq)]
 
-// FIXME: Get rid of as many `pub(super)` declarations as we can.
 // TODO:  Should `Leaf` move into `Tree`/`Configuration` since we might want the tree to have
 //        special kinds of leaf input (metadata along with just the digest)?
 // TODO:  Maybe we should require `INNER_HEIGHT` instead of `HEIGHT` so that we don't have to rely
@@ -36,9 +35,6 @@ pub mod inner_tree;
 pub mod partial;
 pub mod path;
 pub mod single_path;
-
-#[cfg(feature = "test")]
-#[cfg_attr(doc_cfg, doc(cfg(feature = "test")))]
 pub mod test;
 
 pub use node::*;
